requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
python-telegram-bot==20.7
